Move contributed custom editors into common
