+++ src/vs/workbench/contrib/customEditor/common/contributedCustomEditors.ts
